+++ protected/humhub/modules/notification/jobs/SendBulkNotification.php
@@ -8,12 +8,14 @@
 
 namespace humhub\modules\notification\jobs;
 
+use humhub\modules\notification\components\BaseNotification;
 use humhub\modules\queue\LongRunningActiveJob;
 use humhub\modules\user\components\ActiveQueryUser;
+use humhub\modules\user\models\User;
 use Yii;
 
 /**
- * Description of SendNotification
+ * Send Bulk Notification
  *
  * @author buddha
  * @since 1.2
@@ -21,7 +23,7 @@ use Yii;
 class SendBulkNotification extends LongRunningActiveJob
 {
     /**
-     * @var array Basenotification data as array.
+     * @var BaseNotification Base notification object
      */
     public $notification;
 
@@ -31,10 +33,93 @@ class SendBulkNotification extends LongRunningActiveJob
     public $query;
 
     /**
+     * @var string|null $uid Unique ID to cache already processed user IDs
+     * @since 1.17.3
+     */
+    public ?string $uid = null;
+
+    /**
+     * @inheritdoc
+     */
+    public function init()
+    {
+        parent::init();
+
+        if ($this->uid === null) {
+            $this->uid = Yii::$app->security->generateRandomString();
+        }
+    }
+
+    /**
      * @inheritdoc
      */
     public function run()
     {
-        Yii::$app->notification->sendBulk($this->notification, $this->query);
+        Yii::$app->notification->sendBulk($this);
+    }
+
+    /**
+     * Get a cache key for processed user IDs
+     *
+     * @return string
+     * @since 1.17.3
+     */
+    private function getProcessedCacheKey(): string
+    {
+        return 'SendBulkNotification' . $this->uid;
+    }
+
+    /**
+     * Get the array with processed user IDs
+     *
+     * @return array
+     * @since 1.17.3
+     */
+    public function getProcessedUserIds(): array
+    {
+        $processedUserIds = Yii::$app->cache->get($this->getProcessedCacheKey());
+        return is_array($processedUserIds) ? $processedUserIds : [];
+    }
+
+    /**
+     * Check if the user has been already processed
+     *
+     * @param User $user
+     * @return bool
+     * @since 1.17.3
+     */
+    public function isProcessedUser(User $user): bool
+    {
+        return in_array($user->id, $this->getProcessedUserIds());
+    }
+
+    /**
+     * Mark the user as processed to avoid duplicated notifications
+     *
+     * @param User $user
+     * @return void
+     * @since 1.17.3
+     */
+    public function acknowledge(User $user): void
+    {
+        $processedUserIds = $this->getProcessedUserIds();
+        $processedUserIds[] = $user->id;
+        Yii::$app->cache->set($this->getProcessedCacheKey(), $processedUserIds);
+    }
+
+    /**
+     * Get users query with excluding already processed users
+     *
+     * @return ActiveQueryUser
+     * @since 1.17.3
+     */
+    public function getQuery(): ActiveQueryUser
+    {
+        $processedUserIds = $this->getProcessedUserIds();
+        if ($processedUserIds !== []) {
+            $this->query->andWhere(['NOT IN', 'user.id', $processedUserIds]);
+        }
+
+        return $this->query;
     }
 }